#15 Extract serializer and add doc comments
